+++ remember-me-controls.php
@@ -291,10 +291,9 @@ final class c2c_RememberMeControls extends c2c_Plugin_064 {
 	 */
 	public function add_css() {
 		$options = $this->get_options();
-		$type_attr = current_theme_supports( 'html5', 'style' ) ? '' : ' type="text/css"';
 
 		if ( $options['disable_remember_me'] ) {
-			echo '<style' . $type_attr . '>.forgetmenot { display:none; }</style>' . "\n";
+			echo '<style>.forgetmenot { display:none; }</style>' . "\n";
 		}
 	}
 
@@ -303,18 +302,17 @@ final class c2c_RememberMeControls extends c2c_Plugin_064 {
 	 */
 	public function add_js() {
 		$options = $this->get_options();
-		$type_attr = current_theme_supports( 'html5', 'script' ) ? '' : ' type="text/javascript"';
 
 		if ( $options['auto_remember_me'] && ! $options['disable_remember_me'] ) {
-			echo <<<JS
-		<script{$type_attr}>
+			echo <<<HTML
+		<script>
 			const rememberme_checkbox = document.getElementById('rememberme');
 			if ( null !== rememberme_checkbox ) {
 				rememberme_checkbox.checked = true;
 			}
 		</script>
 
-JS;
+HTML;
 		}
 	}